requirements.txt
omxplayer-wrapper
phue
